backend/requirements.txt
fastapi
uvicorn
pymongo
openai
pydantic
python-dotenv
motor
passlib[bcrypt]
python-multipart
jose[cryptography]
